все работает. Перед разбивкой на модули
