+++ src/index.html
@@ -83,7 +83,7 @@
           <p class="nothing-found__mini-text">Ничего не найдено</p>
           <p class="nothing-found__text">К сожалению по вашему запросу ничего не найдено.</p>
         </div>
-        <p class="server__error">Что-то пошло не так =( Похоже, ошибка на севере</p>
+        <p class="server__error">Во время запроса произошла ошибка. Возможно, проблема с соединением или сервер недоступен. Подождите немного и попробуйте ещё раз</p>
       </div>
       <!--об авторе--------------------------------------------------------->
       <div class="cover__about-author">
