+++ src/js/constants/constants.js
@@ -1,8 +1,8 @@
 //newsapi
-const KEY = da509dbac2d74492a4498c044ca28d18;
-const URL_START = 'http://newsapi.org/v2/everything?q=';
-const URL_END = 'sortBy=popularity&apiKey=' + KEY;
-const LANGUAGE = 'language=ru';
+export const URL_START = 'https://newsapi.org/v2/everything?q=';
+export const URL_END = 'sortBy=popularity&pageSize=100&apiKey=da509dbac2d74492a4498c044ca28d18';
+
+export const MS_IN_DAY = 86400000;//или лучше формулой?
 //список карточек
 const NUMBER_OF_CARDS = 3;
 // дата
